requirements.txt
pandas==1.5.0
matplotlib==3.7.0
psycopg2-binary == 2.9.10
aiogram == 3.0.0b7
numpy == 1.24.4
